+++ src/cascade/dismod/__init__.py
@@ -1,2 +1,6 @@
-from cascade.core.log import getLoggers
-CODELOG, MATHLOG = getLoggers(__name__)
+from cascade.dismod.model import Model
+from cascade.dismod.session import Session
+from cascade.dismod.smooth_grid import SmoothGrid
+from cascade.dismod.var import Var
+
+__all__ = [Model, Session, SmoothGrid, Var]
